test(logic): split fog_resolution tests for #2288
Refs #2288

Replace monolithic fog_resolution_test.dart (750 lines) with three
part files under 400 lines each, grouped by concern. No test logic
changes; trim redundant imports where analyzer-clean.
